UI update
Fixed a bug whereby applying customization would open the login menu.

Adjusted the character information screen so that the apply button is in a more reasonable place. Also improved the visual of the Name section by widening the edit box.

Also made the Custom and Sabers buttons a bit prettier in placement.